PolynomialsMutableArithmetics.jl is a Julia package that registers the Polynomial type from the Polynomials package with the MutableArithmetics package. This allows some operations with polynomials to allocate significantly less.

For example, with

using Polynomials
d, m, n = 30, 20, 20
p(d) = Polynomial(big.(1:d))
A = [p(d) for i in 1:m, j in 1:n]
b = [p(d) for i in 1:n]

The multiplication A*b allocates about 40 MiB due to 2 M allocations, as it does not exploit any mutability.

Whereas,

using PolynomialsMutableArithmetics
import MutableArithmetics
MutableArithmetics.operate(*, A, b)

allocates about 70 KiB due to 4k allocations.

PolynomialsMutableArithmetics

Documentation for PolynomialsMutableArithmetics.

PolynomialsMutableArithmetics.add_conv

PolynomialsMutableArithmetics.@register_mutable_arithmetic

PolynomialsMutableArithmetics.add_conv — Function

add_conv(out::Vector{T}, E::Vector{T}, k::Vector{T})

Returns the vector out + fastconv(E, k). Note that only MA.buffered_operate! is implemented.

PolynomialsMutableArithmetics.@register_mutable_arithmetic — Macro

@register_mutable_arithmetic

Register polynomial type (with vector based backend) to work with MutableArithmetics